+++ docker/resource_docker_container.go
@@ -205,12 +205,14 @@ func resourceDockerContainer() *schema.Resource {
 
 						"external": &schema.Schema{
 							Type:     schema.TypeInt,
+							Default:  "32678",
 							Optional: true,
 							ForceNew: true,
 						},
 
 						"ip": &schema.Schema{
 							Type:     schema.TypeString,
+							Default:  "0.0.0.0",
 							Optional: true,
 							ForceNew: true,
 						},