+++ docker/resource_docker_container_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	"context"
+
 	"github.com/docker/docker/api/types"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
@@ -375,6 +376,98 @@ func TestAccDockerContainer_device(t *testing.T) {
 		},
 	})
 }
+func TestAccDockerContainer_port_internal(t *testing.T) {
+	var c types.ContainerJSON
+
+	testCheck := func(*terraform.State) error {
+		portMap := c.NetworkSettings.NetworkSettingsBase.Ports
+		portBindings, ok := portMap["80/tcp"]
+		if !ok || len(portMap["80/tcp"]) == 0 {
+			return fmt.Errorf("Port 80 on tcp is not set")
+		}
+
+		portBindingsLength := len(portBindings)
+		if portBindingsLength != 1 {
+			return fmt.Errorf("Expected 1 binding on port 80, but was %d", portBindingsLength)
+		}
+
+		if len(portBindings[0].HostIP) == 0 {
+			return fmt.Errorf("Expected host IP to be set, but was empty")
+		}
+
+		if len(portBindings[0].HostPort) == 0 {
+			return fmt.Errorf("Expected host port to be set, but was empty")
+		}
+
+		return nil
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccDockerContainerInternalPortConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccContainerRunning("docker_container.foo", &c),
+					testCheck,
+					resource.TestCheckResourceAttr("docker_container.foo", "name", "tf-test"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.#", "1"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.internal", "80"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.protocol", "tcp"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.external", "32678"),
+				),
+			},
+		},
+	})
+}
+func TestAccDockerContainer_port(t *testing.T) {
+	var c types.ContainerJSON
+
+	testCheck := func(*terraform.State) error {
+		portMap := c.NetworkSettings.NetworkSettingsBase.Ports
+		portBindings, ok := portMap["80/tcp"]
+		if !ok || len(portMap["80/tcp"]) == 0 {
+			return fmt.Errorf("Port 80 on tcp is not set")
+		}
+
+		portBindingsLength := len(portBindings)
+		if portBindingsLength != 1 {
+			return fmt.Errorf("Expected 1 binding on port 80, but was %d", portBindingsLength)
+		}
+
+		if len(portBindings[0].HostIP) == 0 {
+			return fmt.Errorf("Expected host IP to be set, but was empty")
+		}
+
+		if len(portBindings[0].HostPort) == 0 {
+			return fmt.Errorf("Expected host port to be set, but was empty")
+		}
+
+		return nil
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccDockerContainerPortConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccContainerRunning("docker_container.foo", &c),
+					testCheck,
+					resource.TestCheckResourceAttr("docker_container.foo", "name", "tf-test"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.#", "1"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.internal", "80"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.protocol", "tcp"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.external", "32787"),
+				),
+			},
+		},
+	})
+}
 
 func testAccContainerRunning(n string, container *types.ContainerJSON) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
@@ -536,6 +629,38 @@ resource "docker_container" "foo" {
 	devices {
     host_path = "/dev/zero"
     container_path = "/dev/zero_test"
+	}
+}
+`
+
+const testAccDockerContainerInternalPortConfig = `
+resource "docker_image" "foo" {
+	name = "nginx:latest"
+	keep_locally = true
+}
+
+resource "docker_container" "foo" {
+	name = "tf-test"
+	image = "${docker_image.foo.latest}"
+	
+	ports {
+		internal = "80"
+	}
+}
+`
+const testAccDockerContainerPortConfig = `
+resource "docker_image" "foo" {
+	name = "nginx:latest"
+	keep_locally = true
+}
+
+resource "docker_container" "foo" {
+	name = "tf-test"
+	image = "${docker_image.foo.latest}"
+
+	ports {
+		internal = "80"
+		external = "32787"
 	}
 }
 `